% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p2.R
\name{fit_p1}
\alias{fit_p1}
\title{Fitting function for p1 models}
\usage{
fit_p1(y, XnS, XnR, XvD, XvC, trace = FALSE, init = NULL, opt = nlminb,
  ...)
}
\arguments{
\item{y}{An adjacency matrix of size \eqn{g \times g}{g x g}.}

\item{XnS}{Matrix of sender effects of size \eqn{g \times k_s}{g x ks}.}

\item{XnR}{Matrix of receiver effects of size \eqn{g \times k_r}{g x kr}.}

\item{XvD}{Density effects, a 3-dim array of size \eqn{g \times g \times
k_d}{g x g x kd}.}

\item{XvC}{Reciprocity effects, a 3-dim array of size \eqn{g \times g \times
k_c}{g x g x kc}.}

\item{trace}{TRUE for tracing information during the estimation. Default is FALSE.}

\item{init}{Optional starting value for model parameters. Default is NULL.}

\item{opt}{Name of the optimising function. Default is nlminb.}

\item{...}{Optional arguments passed to the optimiser.}
}
\value{
The returned value is an object of class
\code{"p1"}, a list  containing the following components:
\item{\code{theta}}{the vector of model estimates.}
\item{\code{loglik}}{the log likelihood function at the estimate.}
\item{\code{AIC, BIC}}{model selection criteria at the estimate.}
\item{\code{XnS.null, XnR.null}}{logical, flagging whether there
  are no sender or receiver effects, respectively.}
\item{\code{seed}}{the random seed used for estimation (when \code{M}>0).}
\item{\code{theta.cov}}{Variance matrix of estimates.}
\item{\code{theta.se}}{Standard errors of estimates.}
\item{\code{opt}}{Optimiser employed for estimation.}
\item{\code{opt.details}}{Object returned by the optimiser.}
\item{\code{ADobj}}{Object returned by \code{TMB:MakeADFun}, containing
the log likelihood function and its gradient (among its components).}
\item{\code{model.data}}{List containing  all the model data, fed
to \code{TMB:MakeADFun}.}
\item{\code{sdrep}}{Summary object returned by \code{TMB::sdreport}. This is an
object with many slots, useful for TMB users.}
}
\description{
Estimates a p1 model via (exact) maximum likelihood.
}
\details{
The function is included for compatibility with \code{fit_p2}, and
it might also be useful to obtain starting values.
}
\examples{
# Analysis of the  kracknets data from the NetData package
library(NetData)
data(kracknets)
# data preparation
g <- 21
Y <- matrix(0, g, g)
ind <-1
for(i in 1:nrow(friendship_data_frame)){
   sele <- friendship_data_frame[i, ]
   Y[sele$ego, sele$alter] <- sele$friendship_tie
   }
Xn <- model.matrix(~ AGE + TENURE, attributes)[, -1]
XvD <- array(1, dim=c(g, g, 4))
for(i in 1:g)
 for(j in 1:g){
    XvD[i, j, 2] <- as.numeric(attributes$DEPT[i]==attributes$DEPT[j])
    XvD[i, j, 3] <- as.numeric(attributes$LEVEL[i]==attributes$LEVEL[j])
    XvD[i, j, 4] <- abs(attributes$AGE[i] - attributes$AGE[j])
  }
XvC <- array(1, dim=c(g, g, 1))
# Now we are ready to fit the model
mod <- fit_p1(Y, Xn, Xn, XvD, XvC)
print(mod)
}
\author{
Ruggero Bellio
}
