% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IS.R
\name{fit_p2_IS}
\alias{fit_p2_IS}
\title{Improve the fit of p2 models}
\usage{
fit_p2_IS(objfit, M = 5000, ncores = 1, init = NULL, init.hess = TRUE,
  seed = NULL, trace = 0, hessian = 1, grtol = 1e-06, xtol = 1e-12)
}
\arguments{
\item{objfit}{An object produced by a call to \code{fit_p2}.}

\item{M}{number of importance sampling simulations. Default is \code{M=5000}.}

\item{ncores}{Number of cores to employ.}

\item{init}{Starting point for the model parameter value. If it is
\code{NULL}, the \code{theta} slot of \code{objfit} is used.}

\item{init.hess}{Should the hessian matrix from \code{fit_p2} be used
as the initial Hessian by the \code{ucminf} optimiser? Default is \code{TRUE}.}

\item{seed}{Random seed for the importance sampler.}

\item{trace, hessian, grtol, xtol}{Arguments for the \code{ucminf} optimiser. For
their meaning, see the help file for that function.}
}
\value{
The returned value is an object of class
\code{"p2"}, a list  containing the same components listed for \code{fit_p2}.
}
\description{
Estimates a p2 model using \emph{parameter dependent Laplace Importance Sampling}.
}
\details{
The function can be called only after fitting a model by \code{fit_p2}, so that
all the model settings are inherited from those employed in the call to\code{fit_p2}.
The function can take advantage of multiple cores for parallel evaluation of the
samples simulated by the importance sampler. The optimiser used is \code{ucminf}, which
can exploit the estimated Hessian from  \code{fit_p2} in a Quasi-Newton optimisation.
}
\examples{
\dontrun{
mod <- fit_p2(Y, Xn, Xn, XvD, XvC)
nc <- detectCores()
modIS <- fit_p2_IS(mod, ncores=nc)}

}
\seealso{
\code{\link{fit_p2}}, \code{\link{ucminf}}
}
\author{
Ruggero Bellio
}
