% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p2.R
\name{gof_p2}
\alias{gof_p2}
\title{Computes GOF statistics for a p2 model fit}
\usage{
gof_p2(objfit, nsim = 100, GOF = ~idegree + odegree + distance,
  verbose = FALSE, conditional = FALSE)
}
\arguments{
\item{objfit}{An object produced by a call to \code{fit_p2}.}

\item{nsim}{Number of simulated networks.}

\item{GOF}{A formula object to specify the statistics to use to diagnosis the goodness-of-fit
of the model, among those suitable for directed networks. See the help file of \code{gof.ergmm}
for more details.}

\item{verbose}{Provides verbose information on the progress of the simulation. Default is FALSE.}

\item{conditional}{Should the random effects be set at the estimated values,
or else simulated from the normal distribution? Default is FALSE.}
}
\value{
Like \code{gof.ergmm}, it returns an object of class \code{gofobject}.
}
\description{
This is a slight modification of the function \code{gof.ergmm} from the
package \code{latentnet}.
}
\details{
The function makes only a minor modification to the function \code{gof.ergmm} from the
package \code{latentnet}. The only change is actually that the function \code{simulate.p2}
is employed to generate the networks.
This means that this function is based on 'statnet' project software
(\url{http://statnet.org}). For license and citation information see \url{http://statnet.org/attribution}.
It would have been much preferable to pass the simulated networks to \code{gof.ergmm}, but
unfortunately such functionality is not supported, therefore it has been necessary to duplicate
the function.
}
