% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p2model.R
\docType{package}
\name{p2model}
\alias{p2model}
\alias{p2model-package}
\title{p2model: A package for fitting p2 models}
\description{
The \code{p2model} package can be used to estimate the parameters of a p2 model
for directed binary networks with correlated random effects. It implements
(approximate) maximum likelihood estimation, following the methodology
studied in Bellio and Soriani (2019).
}
\details{
A vignette is available with some illustrative examples.
}
\references{
Bellio, R. and Soriani, N. (2019). Maximum likelihood estimation based on the
Laplace approximation for \eqn{p_2}{p2} network regression models. \emph{Submitted manuscript}.
}
