% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p2.R
\name{plot_effects_p2}
\alias{plot_effects_p2}
\title{Plot receiver and sender estimated random effects}
\usage{
plot_effects_p2(fit, resolution = round(length(fit$ranef))/2, seed = 2,
  col = rev(grDevices::heat.colors(10)), mai = c(1, 1, 1, 1) * 0.5,
  arrow.size = 0.5, ...)
}
\arguments{
\item{fit}{An object returned by \code{fit_p2}.}

\item{resolution}{Graphical resolution.}

\item{seed}{Random seed.}

\item{col}{Type of colors for the output.}

\item{mai}{Margins for plots.}

\item{arrow.size}{Arrow size.}

\item{...}{Optional graphical arguments.}
}
\description{
Plot receiver and sender estimated random effects
}
