% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p2.R
\name{simulate_p2}
\alias{simulate_p2}
\title{Simulate networks from a p2 model}
\usage{
simulate_p2(objfit, nsim = 100, conditional = FALSE, theta = NULL,
  Sigma = NULL)
}
\arguments{
\item{objfit}{An object produced by a call to \code{fit_p2}.}

\item{nsim}{Number of simulated networks.}

\item{conditional}{Should the random effects be set at the estimated values,
or else simulated from the normal distribution? Default is FALSE.}

\item{theta}{Parameter value to be used for the simulation. In case it is \code{NULL},
the \code{theta} slot of \code{objfit} is used.}

\item{Sigma}{Variance matrix of random effects to be used for the simulation. In case
it is \code{NULL}, the \code{Sigma} slot of \code{objfit} is used.}
}
\value{
Either a single network (when \code{nsim}=1) or a list of \code{nsim} networks. See \code{\link[network]{network}}.
}
\description{
The function can be used to simulate from an estimated model or from a given
parameter value. In the latter case, the model settings are included in a fitted model
object.
}
